fix(website): remove analytics consent popup
