+++ changelog.d/website-remove-analytics-popup.md
@@ -1,0 +1,2 @@
+- **Website analytics popup removed.** Google Analytics now starts automatically
+  on the public documentation site, with no consent popup or preference button.
