+++ website/.vitepress/theme/analytics.mjs
@@ -1,6 +1,5 @@
 // Public documentation only. The apps and embedded Studio never import this.
 export const measurementId = 'G-RG6PPTGX2T'
-const storageKey = 'mold.website.analytics'
 
 function cleanUrl(value) {
   try {
@@ -12,27 +11,16 @@ function cleanUrl(value) {
 }
 
 export function createAnalytics(win) {
-  let selected
   let loaded = false
   let previousPage = cleanUrl(win.document.referrer)
   const eligible = () =>
     win.location.origin === 'https://utensils.io' &&
     win.location.pathname.startsWith('/mold/')
-  function choice() {
-    if (selected !== undefined) return selected
-    try {
-      const saved = win.localStorage.getItem(storageKey)
-      if (saved === 'granted' || saved === 'denied') selected = saved
-    } catch {
-      // Storage may be disabled. Consent still works for the current page.
-    }
-    return selected
-  }
   function gtag() {
     win.dataLayer.push(arguments)
   }
   function pageView() {
-    if (!eligible() || choice() !== 'granted' || !loaded) return
+    if (!eligible() || !loaded) return
     const page = cleanUrl(win.location.href)
     gtag('set', { page_location: page, page_referrer: previousPage })
     gtag('event', 'page_view', {
@@ -43,7 +31,7 @@ export function createAnalytics(win) {
     previousPage = page
   }
   function start() {
-    if (!eligible() || choice() !== 'granted' || loaded) return
+    if (!eligible() || loaded) return
     loaded = true
     win.dataLayer = win.dataLayer || []
     win.gtag = gtag
@@ -73,19 +61,5 @@ export function createAnalytics(win) {
     win.document.head.appendChild(script)
     pageView()
   }
-  function choose(accepted) {
-    selected = accepted ? 'granted' : 'denied'
-    try {
-      win.localStorage.setItem(storageKey, selected)
-    } catch {
-      // Keep the choice in memory if persistent storage is unavailable.
-    }
-    if (loaded) {
-      win[`ga-disable-${measurementId}`] = !accepted
-      gtag('consent', 'update', { analytics_storage: selected })
-    } else if (accepted) {
-      start()
-    }
-  }
-  return { choice, choose, start, pageView }
+  return { start, pageView }
 }
